requirements.txt
google-generativeai>=0.7.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.5
tqdm>=4.66.0
pandas>=2.1.0
